requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-openai
streamlit
requests
python-multipart
langchain
openai
faiss-cpu
tiktoken
PyMuPDF
docx2txt
python-docx
pandas
pypdf
